+++ src/main/java/woowacourse/shoppingcart/dto/product/PageRequest.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.dto;
+package woowacourse.shoppingcart.dto.product;
 
 import javax.validation.constraints.Min;
